Avancement

+++ app/src/Controller/AdressController.php
@@ -12,11 +12,13 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 #[Route('/adress')]
 final class AdressController extends AbstractController
 {
     #[Route(name: 'app_adress_index', methods: ['GET'])]
+    #[isGranted('ROLE_CUSTOMER')]
     public function index(AdressRepository $adressRepository): Response
     {
         return $this->render('adress/index.html.twig', [
@@ -25,6 +27,7 @@ final class AdressController extends AbstractController
     }
 
     #[Route('/new', name: 'app_adress_new', methods: ['GET', 'POST'])]
+    #[isGranted('ROLE_CUSTOMER')]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
         $adress = new Adress();
@@ -45,16 +48,28 @@ final class AdressController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_adress_show', methods: ['GET'])]
+    #[isGranted('ROLE_CUSTOMER')]
     public function show(Adress $adress): Response
     {
+        $user = $this->getUser();
+        if ($adress->getUsers() !== $user) {
+            $this->addFlash('danger', 'Vous n\'avez pas le droit' );
+            return $this->redirectToRoute('app_adress_index', [], Response::HTTP_SEE_OTHER);
+        }
         return $this->render('adress/show.html.twig', [
             'adress' => $adress,
         ]);
     }
 
     #[Route('/{id}/edit', name: 'app_adress_edit', methods: ['GET', 'POST'])]
+    #[isGranted('ROLE_CUSTOMER')]
     public function edit(Request $request, Adress $adress, EntityManagerInterface $entityManager): Response
     {
+        $user = $this->getUser();
+        if ($adress->getUsers() !== $user) {
+            $this->addFlash('danger', 'Vous n\'avez pas le droit' );
+            return $this->redirectToRoute('app_adress_index', [], Response::HTTP_SEE_OTHER);
+        }
         $form = $this->createForm(AdressForm::class, $adress);
         $form->handleRequest($request);
 
@@ -71,8 +86,15 @@ final class AdressController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_adress_delete', methods: ['POST'])]
+    #[isGranted('ROLE_CUSTOMER')]
     public function delete(Request $request, Adress $adress, EntityManagerInterface $entityManager): Response
     {
+        $user = $this->getUser();
+
+        if ($adress->getUsers() !== $user) {
+            $this->addFlash('danger', 'Vous n\'avez pas le droit' );
+            return $this->redirectToRoute('app_adress_index', [], Response::HTTP_SEE_OTHER);
+        }
         if ($this->isCsrfTokenValid('delete'.$adress->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($adress);
             $entityManager->flush();
@@ -82,6 +104,7 @@ final class AdressController extends AbstractController
     }
 
     #[Route('/{id}/favorite', name: 'address_set_favorite', methods: ['POST'])]
+    #[isGranted('ROLE_CUSTOMER')]
     public function setFavorite(
         Adress $address,
         AddressManager $addressManager,
